Move menu loader file to parent folder
