+++ www/front_src/src/components/MenuLoader.tsx
