+++ www/front_src/src/components/header/hostMenu/index.js
@@ -39,7 +39,7 @@ import {
   pendingCriterias,
   unhandledStateCriterias,
 } from '../getResourcesUrl';
-import MenuLoader from '../menuLoader';
+import MenuLoader from '../../MenuLoader';
 
 const numberFormat = yup.number().required().integer();
